+++ cmd/agent.go
@@ -305,6 +305,11 @@ var agentCreateCmd = &cobra.Command{
 				errsystem.New(errsystem.ErrInvalidConfiguration, err, errsystem.WithAttributes(map[string]any{"identifier": theproject.Project.Bundler.Identifier})).ShowErrorAndExit()
 			}
 
+			template, err := templates.LoadTemplateForRuntime(context.Background(), tmpdir, theproject.Project.Bundler.Identifier)
+			if err != nil {
+				errsystem.New(errsystem.ErrInvalidConfiguration, err, errsystem.WithAttributes(map[string]any{"identifier": theproject.Project.Bundler.Identifier})).ShowErrorAndExit()
+			}
+
 			if err := rules.NewAgent(templates.TemplateContext{
 				Logger:           logger,
 				AgentName:        name,
@@ -313,6 +318,7 @@ var agentCreateCmd = &cobra.Command{
 				AgentDescription: description,
 				ProjectDir:       theproject.Dir,
 				TemplateDir:      tmpdir,
+				Template:         template,
 				AgentuityCommand: getAgentuityCommand(),
 			}); err != nil {
 				errsystem.New(errsystem.ErrApiRequest, err, errsystem.WithAttributes(map[string]any{"name": name})).ShowErrorAndExit()
@@ -409,29 +415,29 @@ func reconcileAgentList(logger logger.Logger, cmd *cobra.Command, apiUrl string,
 	for _, filename := range localAgents {
 		agentName := filepath.Base(filepath.Dir(filename))
 		key := normalAgentName(agentName)
-		var found bool
-		for _, agent := range remoteAgents {
-			if localAgent, ok := fileAgentsByID[agent.ID]; ok {
-				if localAgent.Name == agentName {
-					oldkey := normalAgentName(agent.Name)
-					agent.Name = localAgent.Name
-					state[key] = agentListState{
-						Agent:       &agent,
-						Filename:    filename,
-						FoundLocal:  true,
-						FoundRemote: true,
-						Rename:      true,
-						RenameFrom:  oldkey,
-					}
-					delete(state, oldkey)
-					found = true
-					break
-				}
-			}
-		}
-		if found {
-			continue
-		}
+		// var found bool
+		// for _, agent := range remoteAgents {
+		// 	if localAgent, ok := fileAgentsByID[agent.ID]; ok {
+		// 		if localAgent.Name == agentName {
+		// 			oldkey := normalAgentName(agent.Name)
+		// 			agent.Name = localAgent.Name
+		// 			state[key] = agentListState{
+		// 				Agent:       &agent,
+		// 				Filename:    filename,
+		// 				FoundLocal:  true,
+		// 				FoundRemote: true,
+		// 				Rename:      true,
+		// 				RenameFrom:  oldkey,
+		// 			}
+		// 			delete(state, oldkey)
+		// 			found = true
+		// 			break
+		// 		}
+		// 	}
+		// }
+		// if found {
+		// 	continue
+		// }
 		if filepath.Base(filename) == agentFilename {
 			if found, ok := state[key]; ok {
 				state[key] = agentListState{