+++ cmd/cloud.go
@@ -23,6 +23,7 @@ import (
 	"github.com/agentuity/go-common/crypto"
 	"github.com/agentuity/go-common/env"
 	"github.com/agentuity/go-common/logger"
+	cstr "github.com/agentuity/go-common/string"
 	"github.com/agentuity/go-common/tui"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -90,7 +91,7 @@ func ShowNewProjectImport(ctx context.Context, logger logger.Logger, cmd *cobra.
 	tui.WaitForAnyKey()
 	tui.ClearScreen()
 	orgId := promptForOrganization(ctx, logger, cmd, apiUrl, apikey)
-	name, description := promptForProjectDetail(ctx, logger, apiUrl, apikey, project.Name, project.Description)
+	name, description := promptForProjectDetail(ctx, logger, apiUrl, apikey, project.Name, project.Description, orgId)
 	project.Name = name
 	project.Description = description
 	var createWebhookAuth bool
@@ -146,6 +147,12 @@ Examples:
 		transportUrl := context.TransportURL
 		token := context.Token
 		ci, _ := cmd.Flags().GetBool("ci")
+		ciRemoteUrl, _ := cmd.Flags().GetString("ci-remote-url")
+		ciBranch, _ := cmd.Flags().GetString("ci-branch")
+		ciCommit, _ := cmd.Flags().GetString("ci-commit")
+		ciMessage, _ := cmd.Flags().GetString("ci-message")
+		ciGitProvider, _ := cmd.Flags().GetString("ci-git-provider")
+		ciLogsUrl, _ := cmd.Flags().GetString("ci-logs-url")
 
 		deploymentConfig := project.NewDeploymentConfig()
 		client := util.NewAPIClient(ctx, logger, apiUrl, token)
@@ -221,7 +228,7 @@ Examples:
 					if projectData != nil && projectData.Env != nil && projectData.Env[ev.Key] == ev.Val {
 						continue
 					}
-					if projectData != nil && projectData.Secrets != nil && projectData.Secrets[ev.Key] == ev.Val {
+					if projectData != nil && projectData.Secrets != nil && projectData.Secrets[ev.Key] == cstr.Mask(ev.Val) {
 						continue
 					}
 					foundkeys = append(foundkeys, ev.Key)
@@ -353,24 +360,43 @@ Examples:
 			deploymentId = "/" + deploymentId
 		}
 
-		gitInfo, err := deployer.GetGitInfo(dir)
-		if err != nil {
-			logger.Error("error getting git info: %s", err)
-		}
+		var gitInfo deployer.GitInfo
 		var originType string
-		if ci {
+		var ciInfo deployer.CIInfo
+		isOverwritingGitInfo := ciRemoteUrl != "" || ciBranch != "" || ciCommit != "" || ciMessage != "" || ciGitProvider != "" || ciLogsUrl != ""
+		if ci && isOverwritingGitInfo {
 			originType = "ci"
+			ciInfo = deployer.CIInfo{
+				LogsURL: ciLogsUrl,
+			}
+			gitInfo = deployer.GitInfo{
+				RemoteURL:     &ciRemoteUrl,
+				Branch:        &ciBranch,
+				Commit:        &ciCommit,
+				CommitMessage: &ciMessage,
+				GitProvider:   &ciGitProvider,
+				IsRepo:        true,
+			}
 		} else {
+			info, err := deployer.GetGitInfoRecursive(logger, dir)
+			if err != nil {
+				logger.Debug("Failed to get git info: %v", err)
+			}
+			gitInfo = *info
 			originType = "cli"
 		}
 
+		data := map[string]interface{}{
+			"machine": deployer.GetMachineInfo(),
+			"git":     gitInfo,
+		}
+		if originType == "ci" && ciLogsUrl != "" {
+			data["ci"] = ciInfo
+		}
 		startRequest.Metadata = &deployer.Metadata{
 			Origin: deployer.MetadataOrigin{
 				Type: originType,
-				Data: map[string]interface{}{
-					"machine": deployer.GetMachineInfo(),
-					"git":     gitInfo,
-				},
+				Data: data,
 			},
 		}
 
@@ -632,8 +658,22 @@ func init() {
 	cloudDeployCmd.Flags().StringP("dir", "d", ".", "The directory to the project to deploy")
 	cloudDeployCmd.Flags().String("deploymentId", "", "Used to track a specific deployment")
 	cloudDeployCmd.Flags().Bool("ci", false, "Used to track a specific CI job")
+	cloudDeployCmd.Flags().String("ci-remote-url", "", "Used to set the remote repository URL for your deployment metadata")
+	cloudDeployCmd.Flags().String("ci-branch", "", "Used to set the branch name for your deployment metadata")
+	cloudDeployCmd.Flags().String("ci-commit", "", "Used to set the commit hash for your deployment metadata")
+	cloudDeployCmd.Flags().String("ci-message", "", "Used to set the commit message for your deployment metadata")
+	cloudDeployCmd.Flags().String("ci-git-provider", "", "Used to set the git provider for your deployment metadata")
+	cloudDeployCmd.Flags().String("ci-logs-url", "", "Used to set the CI logs URL for your deployment metadata")
+
 	cloudDeployCmd.Flags().MarkHidden("deploymentId")
 	cloudDeployCmd.Flags().MarkHidden("ci")
+	cloudDeployCmd.Flags().MarkHidden("ci-remote-url")
+	cloudDeployCmd.Flags().MarkHidden("ci-branch")
+	cloudDeployCmd.Flags().MarkHidden("ci-commit")
+	cloudDeployCmd.Flags().MarkHidden("ci-message")
+	cloudDeployCmd.Flags().MarkHidden("ci-git-provider")
+	cloudDeployCmd.Flags().MarkHidden("ci-logs-url")
+
 	cloudDeployCmd.Flags().String("format", "text", "The output format to use for results which can be either 'text' or 'json'")
 	cloudDeployCmd.Flags().String("org-id", "", "The organization to create the project in")
 	cloudDeployCmd.Flags().String("templates-dir", "", "The directory to load the templates. Defaults to loading them from the github.com/agentuity/templates repository")